refactor: enforce max-lines: 200 via oxlint; split 9 oversized source files
Adds `max-lines: ["error", 200]` to `.oxlintrc.json` (tests exempted via
overrides). Splits 9 non-test source files that exceeded the limit into
focused sub-modules; preserves every existing import path with barrel
re-exports so no test changes are required.

Split summary:
- src/types.mts → types/{github,report,iterate}.mts (barrel)
- src/config/load.mts → config/compat.mts (applyCompat extracted)
- src/cli/args.mts → cli/exit-codes.mts (exit-code helpers extracted)
- src/commands/iterate.mts → commands/iterate/{index,helpers,classify,render,stall,escalate,fix-code,steps}.mts (barrel)
- src/commands/check.mts → commands/check-status.mts (computeStatus extracted)
- src/commands/resolve.mts → commands/resolve-instructions.mts (buildFetchInstructions extracted)
- src/github/batch.mts → github/batch-{parsers,raw-types}.mts
- src/cli-parser.mts → cli/{formatters,handlers,iterate-formatter,fix-formatter}.mts
- src/commands/commit-suggestion.mts → trimmed docstring to fit
