+++ .oxlintrc.json
@@ -2,6 +2,15 @@
   "$schema": "https://raw.githubusercontent.com/oxc-project/oxlint/main/npm/oxlint/configuration_schema.json",
   "rules": {
     "unicorn/prefer-node-protocol": "error",
-    "radix": "error"
-  }
+    "radix": "error",
+    "max-lines": ["error", 200]
+  },
+  "overrides": [
+    {
+      "files": ["*.test.mts", "*.mock.test.mts"],
+      "rules": {
+        "max-lines": "off"
+      }
+    }
+  ]
 }
