+++ src/cli-parser.mts
@@ -19,25 +19,24 @@ import { readFileSync } from "node:fs";
 
 import { runCheck } from "./commands/check.mts";
 import { runResolveFetch, runResolveMutate } from "./commands/resolve.mts";
-import { runCommitSuggestion } from "./commands/commit-suggestion.mts";
-import { runIterate, renderResolveCommand } from "./commands/iterate.mts";
-import { runStatus, formatStatusTable } from "./commands/status.mts";
-import { getRepoInfo } from "./github/client.mts";
 import { formatJson } from "./reporters/json.mts";
 import { formatText } from "./reporters/text.mts";
-import { loadConfig } from "./config/load.mts";
 import {
   parseCommonArgs,
   getFlag,
   hasFlag,
   parseList,
-  parseStatusPrNumbers,
-  parseDurationToMinutes,
-  parseIntStrict,
   statusToExitCode,
-  iterateActionToExitCode,
-  deriveSimpleReady,
 } from "./cli/args.mts";
+import {
+  formatFetchResult,
+  formatMutateResult,
+} from "./cli/formatters.mts";
+import {
+  handleCommitSuggestion,
+  handleIterate,
+  handleStatus,
+} from "./cli/handlers.mts";
 
 // ---------------------------------------------------------------------------
 // Entry
@@ -98,7 +97,6 @@ async function handleCheck(args: string[]): Promise<void> {
   process.stdout.write(`${output}\n`);
 
   process.exitCode = statusToExitCode(report.status);
-  return;
 }
 
 async function handleResolve(args: string[]): Promise<void> {
@@ -138,460 +136,4 @@ async function handleResolve(args: string[]): Promise<void> {
         : formatMutateResult(result),
     );
   }
-}
-
-async function handleCommitSuggestion(args: string[]): Promise<void> {
-  const { prNumber, global: globalOpts, extra } = parseCommonArgs(args);
-
-  const threadId = getFlag(extra, "--thread-id");
-  if (!threadId) {
-    process.stderr.write(
-      "Usage: pr-shepherd commit-suggestion [PR] --thread-id ID --message MSG [--description DESC]\n",
-    );
-    process.exitCode = 1;
-    return;
-  }
-
-  const message = getFlag(extra, "--message");
-  if (!message || message.trim() === "") {
-    process.stderr.write("--message is required and must be non-empty\n");
-    process.exitCode = 1;
-    return;
-  }
-
-  const description = getFlag(extra, "--description") ?? undefined;
-
-  const result = await runCommitSuggestion({
-    ...globalOpts,
-    prNumber,
-    threadId,
-    message,
-    description,
-  });
-
-  process.stdout.write(
-    globalOpts.format === "json"
-      ? `${JSON.stringify(result, null, 2)}\n`
-      : formatCommitSuggestionResult(result),
-  );
-
-  process.exitCode = result.applied ? 0 : 1;
-}
-
-async function handleIterate(args: string[]): Promise<void> {
-  const { prNumber, global: globalOpts, extra } = parseCommonArgs(args);
-
-  const lastPushTimeStr = getFlag(extra, "--last-push-time");
-  const lastPushTime = lastPushTimeStr
-    ? parseIntStrict(lastPushTimeStr, "--last-push-time")
-    : undefined;
-  const readyDelayStr = getFlag(extra, "--ready-delay");
-  const cfg = loadConfig();
-  const readyDelaySeconds =
-    parseDurationToMinutes(readyDelayStr ?? "", cfg.watch.readyDelayMinutes) * 60;
-  const cooldownSecondsStr = getFlag(extra, "--cooldown-seconds");
-  const cooldownSeconds = cooldownSecondsStr
-    ? parseIntStrict(cooldownSecondsStr, "--cooldown-seconds")
-    : cfg.iterate.cooldownSeconds;
-  const noAutoMarkReady = hasFlag(extra, "--no-auto-mark-ready");
-  const noAutoCancelActionable = hasFlag(extra, "--no-auto-cancel-actionable");
-  const stallTimeoutStr = getFlag(extra, "--stall-timeout");
-  const stallTimeoutSeconds = stallTimeoutStr
-    ? parseDurationToMinutes(stallTimeoutStr, cfg.iterate.stallTimeoutMinutes) * 60
-    : cfg.iterate.stallTimeoutMinutes * 60;
-
-  const result = await runIterate({
-    ...globalOpts,
-    prNumber,
-    lastPushTime,
-    readyDelaySeconds,
-    cooldownSeconds,
-    stallTimeoutSeconds,
-    noAutoMarkReady,
-    noAutoCancelActionable,
-  });
-
-  if (globalOpts.format === "json") {
-    process.stdout.write(`${JSON.stringify(result)}\n`);
-  } else {
-    process.stdout.write(`${formatIterateResult(result)}\n`);
-  }
-
-  process.exitCode = iterateActionToExitCode(result.action);
-  return;
-}
-
-async function handleStatus(args: string[]): Promise<void> {
-  const { global: globalOpts } = parseCommonArgs(args);
-
-  const prNumbers = parseStatusPrNumbers(args);
-
-  if (prNumbers.length === 0) {
-    process.stderr.write("Usage: pr-shepherd status PR1 [PR2 …]\n");
-    process.exitCode = 1;
-    return;
-  }
-
-  const repo = await getRepoInfo();
-  const summaries = await runStatus({ ...globalOpts, prNumbers });
-  const output =
-    globalOpts.format === "json"
-      ? JSON.stringify(summaries, null, 2)
-      : formatStatusTable(summaries, `${repo.owner}/${repo.name}`);
-
-  process.stdout.write(`${output}\n`);
-
-  const allReady = summaries.every((s) => deriveSimpleReady(s));
-  process.exitCode = allReady ? 0 : 1;
-  return;
-}
-
-// ---------------------------------------------------------------------------
-// Output formatters
-// ---------------------------------------------------------------------------
-
-function formatFetchResult(result: Awaited<ReturnType<typeof runResolveFetch>>): string {
-  const total =
-    result.actionableThreads.length +
-    result.actionableComments.length +
-    result.changesRequestedReviews.length +
-    result.reviewSummaries.length;
-
-  const sections: string[] = [];
-
-  sections.push(
-    `# PR #${result.prNumber} — Resolve fetch (${total === 0 ? "0 actionable" : `${total} actionable`})`,
-  );
-
-  if (result.actionableThreads.length > 0) {
-    sections.push(
-      `## Actionable Review Threads (${result.actionableThreads.length})` +
-        (result.commitSuggestionsEnabled ? " [commit-suggestions: enabled]" : ""),
-    );
-    sections.push(
-      result.actionableThreads
-        .map((t) => {
-          const suggestionMarker = t.suggestion ? " [suggestion]" : "";
-          return `- \`threadId=${t.id}\` \`${t.path ?? ""}:${t.line ?? "?"}\` (@${t.author})${suggestionMarker}: ${t.body.split("\n")[0]?.slice(0, 100) ?? ""}`;
-        })
-        .join("\n"),
-    );
-  }
-
-  if (result.actionableComments.length > 0) {
-    sections.push(`## Actionable PR Comments (${result.actionableComments.length})`);
-    sections.push(
-      result.actionableComments
-        .map(
-          (c) =>
-            `- \`commentId=${c.id}\` (@${c.author}): ${c.body.split("\n")[0]?.slice(0, 100) ?? ""}`,
-        )
-        .join("\n"),
-    );
-  }
-
-  if (result.changesRequestedReviews.length > 0) {
-    sections.push(
-      `## Pending CHANGES_REQUESTED reviews (${result.changesRequestedReviews.length})`,
-    );
-    sections.push(
-      result.changesRequestedReviews.map((r) => `- \`reviewId=${r.id}\` (@${r.author})`).join("\n"),
-    );
-  }
-
-  if (result.reviewSummaries.length > 0) {
-    sections.push(`## Review summaries (${result.reviewSummaries.length})`);
-    sections.push(
-      result.reviewSummaries
-        .map(
-          (r) => `- \`reviewId=${r.id}\` (@${r.author}): ${r.body.split("\n")[0]!.slice(0, 100)}`,
-        )
-        .join("\n"),
-    );
-  }
-
-  sections.push("## Summary");
-  sections.push(
-    total === 0 ? "0 actionable — all threads resolved/minimized" : `${total} actionable item(s)`,
-  );
-
-  sections.push("## Instructions");
-  sections.push(result.instructions.map((inst, i) => `${i + 1}. ${inst}`).join("\n"));
-
-  return `${sections.join("\n\n")}\n`;
-}
-
-function formatCommitSuggestionResult(
-  result: Awaited<ReturnType<typeof runCommitSuggestion>>,
-): string {
-  const lines: string[] = [];
-  if (result.applied) {
-    lines.push(`Applied suggestion from @${result.author}:`);
-    const range =
-      result.startLine === result.endLine
-        ? `line ${result.startLine}`
-        : `lines ${result.startLine}-${result.endLine}`;
-    lines.push(`  ${result.path} (${range})`);
-    if (result.commitSha) lines.push(`Commit: ${result.commitSha}`);
-    if (result.patch) {
-      lines.push("");
-      lines.push("```diff");
-      lines.push(result.patch.trimEnd());
-      lines.push("```");
-    }
-  } else {
-    lines.push(`Failed to apply suggestion ${result.threadId}:`);
-    lines.push(`  path: ${result.path} (lines ${result.startLine}-${result.endLine})`);
-    lines.push(`  author: @${result.author}`);
-    lines.push(`  reason: ${result.reason ?? "unknown"}`);
-    if (result.patch) {
-      lines.push("");
-      lines.push("```diff");
-      lines.push(result.patch.trimEnd());
-      lines.push("```");
-    }
-  }
-  if (result.postActionInstruction) {
-    lines.push("");
-    lines.push(result.postActionInstruction);
-  }
-  return `${lines.join("\n")}\n`;
-}
-
-function formatMutateResult(result: Awaited<ReturnType<typeof runResolveMutate>>): string {
-  const lines: string[] = [];
-  if (result.resolvedThreads.length)
-    lines.push(
-      `Resolved threads (${result.resolvedThreads.length}): ${result.resolvedThreads.join(", ")}`,
-    );
-  if (result.minimizedComments.length)
-    lines.push(
-      `Minimized comments (${result.minimizedComments.length}): ${result.minimizedComments.join(", ")}`,
-    );
-  if (result.dismissedReviews.length)
-    lines.push(
-      `Dismissed reviews (${result.dismissedReviews.length}): ${result.dismissedReviews.join(", ")}`,
-    );
-  if (result.errors.length) lines.push(`Errors:\n  ${result.errors.join("\n  ")}`);
-  return `${lines.join("\n")}\n`;
-}
-
-/**
- * Format an IterateResult as human-readable Markdown.
- *
- * Load-bearing conventions the monitor SKILL relies on:
- *   1. The H1 heading on line 1 contains `[<ACTION>]` — the action tag identifies
- *      the output for logging and validation. Behavior is driven by `## Instructions`,
- *      not by dispatching on the tag.
- *   2. `[FIX_CODE]` uses the `rebase-and-push` variant: the `resolve` bullet under
- *      `## Post-fix push` wraps the resolve command in backticks — the SKILL
- *      extracts the backticked content for execution.
- *   3. The shell script under `[REBASE]` is inside a ```bash fenced block.
- *   4. Every action ends with a `## Instructions` section — numbered `1.`, `2.`, … —
- *      that tells the monitor exactly what to do with this output. The section is
- *      unconditional: every action, every variant, always emits at least one step.
- *      The SKILL simply follows those steps; it does not need its own dispatch table.
- */
-function formatIterateResult(result: import("./types.mts").IterateResult): string {
-  const heading = `# PR #${result.pr} [${result.action.toUpperCase()}]`;
-  const baseLine = `**status** \`${result.status}\` · **merge** \`${result.mergeStateStatus}\` · **state** \`${result.state}\` · **repo** \`${result.repo}\``;
-  const summaryLine = `**summary** ${result.summary.passing} passing, ${result.summary.skipped} skipped, ${result.summary.filtered} filtered, ${result.summary.inProgress} inProgress · **remainingSeconds** ${result.remainingSeconds} · **copilotReviewInProgress** ${result.copilotReviewInProgress} · **isDraft** ${result.isDraft} · **shouldCancel** ${result.shouldCancel}`;
-  const header = [heading, "", baseLine, summaryLine].join("\n");
-  const checksSection = formatChecksSection(result.checks);
-
-  switch (result.action) {
-    case "cooldown":
-      return [
-        header,
-        "",
-        result.log,
-        "",
-        "## Instructions",
-        "",
-        "1. End this iteration — the next cron fire will recheck once CI starts reporting.",
-      ].join("\n");
-
-    case "wait": {
-      const parts = [header];
-      if (checksSection) parts.push(checksSection);
-      parts.push(
-        result.log,
-        "## Instructions",
-        "1. End this iteration — the next cron fire will recheck.",
-      );
-      return parts.join("\n\n").replace(/\n\n\n+/g, "\n\n");
-    }
-
-    case "rerun_ci": {
-      const rerunInstructions = result.reran.map(
-        (r, i) => `${i + 1}. Run: \`gh run rerun ${r.runId} --failed\``,
-      );
-      rerunInstructions.push(
-        `${rerunInstructions.length + 1}. End this iteration — wait for CI to report results after the re-run.`,
-      );
-      const parts = [header];
-      if (checksSection) parts.push(checksSection);
-      parts.push(result.log, "## Instructions", rerunInstructions.join("\n"));
-      return parts.join("\n\n").replace(/\n\n\n+/g, "\n\n");
-    }
-
-    case "mark_ready": {
-      const parts = [header];
-      if (checksSection) parts.push(checksSection);
-      parts.push(
-        result.log,
-        "## Instructions",
-        "1. The CLI already marked the PR ready for review — end this iteration.",
-      );
-      return parts.join("\n\n").replace(/\n\n\n+/g, "\n\n");
-    }
-
-    case "cancel": {
-      const parts = [header];
-      if (checksSection) parts.push(checksSection);
-      parts.push(
-        result.log,
-        "## Instructions",
-        "1. Invoke `/loop cancel` via the Skill tool.\n2. Stop.",
-      );
-      return parts.join("\n\n").replace(/\n\n\n+/g, "\n\n");
-    }
-
-    case "rebase": {
-      const parts = [header];
-      if (checksSection) parts.push(checksSection);
-      parts.push(
-        result.rebase.reason,
-        "```bash\n" + result.rebase.shellScript + "\n```",
-        "## Instructions",
-        "1. Copy the shell script from the ` ```bash ` block above and run it in Bash.\n2. End this iteration — the next cron fire will check CI after the rebase.",
-      );
-      return parts.join("\n\n").replace(/\n\n\n+/g, "\n\n");
-    }
-
-    case "escalate": {
-      const parts = [header];
-      if (checksSection) parts.push(checksSection);
-      parts.push(
-        result.escalate.humanMessage,
-        "## Instructions",
-        "1. Invoke `/loop cancel` via the Skill tool.\n2. Stop — the PR needs human direction before monitoring can resume.",
-      );
-      return parts.join("\n\n").replace(/\n\n\n+/g, "\n\n");
-    }
-
-    case "fix_code":
-      return formatFixCodeResult(header, checksSection, result);
-  }
-}
-
-function formatChecksSection(checks: import("./types.mts").RelevantCheck[]): string | null {
-  if (checks.length === 0) return null;
-  const lines: string[] = ["## Checks", ""];
-  for (const c of checks) {
-    if (c.conclusion === "SUCCESS") {
-      lines.push(`- ✓ \`${c.name}\` — SUCCESS`);
-    } else {
-      const kind = c.failureKind ? ` (${c.failureKind})` : "";
-      const locator = c.runId
-        ? `\`${c.runId}\``
-        : c.detailsUrl
-          ? `external \`${c.detailsUrl}\``
-          : "(no runId)";
-      lines.push(`- ✗ \`${c.name}\`${kind} — ${c.conclusion} · ${locator}`);
-      if (c.errorExcerpt) {
-        for (const eLine of c.errorExcerpt.split("\n")) {
-          lines.push(`  > ${eLine}`);
-        }
-      }
-    }
-  }
-  return lines.join("\n");
-}
-
-function formatFixCodeResult(
-  header: string,
-  checksSection: string | null,
-  result: import("./types.mts").IterateResultFixCode,
-): string {
-  const sections: string[] = [header];
-  if (checksSection) sections.push(checksSection);
-
-  if (result.fix.threads.length > 0) {
-    sections.push("## Review threads");
-    for (const t of result.fix.threads) {
-      const loc = t.path ? `\`${t.path}:${t.line ?? "?"}\`` : "(no location)";
-      sections.push(`### \`${t.id}\` — ${loc} (@${t.author})`);
-      sections.push(blockquote(t.body));
-    }
-  }
-
-  if (result.fix.actionableComments.length > 0) {
-    sections.push("## Actionable comments");
-    for (const c of result.fix.actionableComments) {
-      sections.push(`### \`${c.id}\` (@${c.author})`);
-      sections.push(blockquote(c.body));
-    }
-  }
-
-  if (result.fix.checks.length > 0) {
-    sections.push("## Failing checks");
-    const bullets = result.fix.checks.map((ch) => {
-      const kind = ch.failureKind ?? "actionable";
-      if (ch.runId) return `- \`${ch.runId}\` — \`${ch.name}\` (${kind})`;
-      if (ch.detailsUrl) return `- external \`${ch.detailsUrl}\` — \`${ch.name}\` (${kind})`;
-      return `- (no runId) — \`${ch.name}\` (${kind})`;
-    });
-    sections.push(bullets.join("\n"));
-  }
-
-  if (result.fix.changesRequestedReviews.length > 0) {
-    sections.push("## Changes-requested reviews");
-    sections.push(
-      result.fix.changesRequestedReviews.map((r) => `- \`${r.id}\` (@${r.author})`).join("\n"),
-    );
-  }
-
-  if (result.fix.noiseCommentIds.length > 0) {
-    sections.push("## Noise (minimize only)");
-    sections.push(result.fix.noiseCommentIds.map((id) => `\`${id}\``).join(", "));
-  }
-
-  if (result.fix.reviewSummaryIds.length > 0) {
-    sections.push("## Review summaries (minimize only)");
-    sections.push(result.fix.reviewSummaryIds.map((id) => `\`${id}\``).join(", "));
-  }
-
-  if (result.fix.surfacedSummaries.length > 0) {
-    sections.push("## Review summaries (surfaced — not minimized)");
-    for (const r of result.fix.surfacedSummaries) {
-      sections.push(`### \`${r.id}\` (@${r.author})`);
-      sections.push(blockquote(r.body));
-    }
-  }
-
-  if (result.cancelled.length > 0) {
-    sections.push("## Cancelled runs");
-    sections.push(result.cancelled.map((id) => `\`${id}\``).join(", "));
-  }
-
-  sections.push("## Post-fix push");
-  sections.push(
-    [
-      `- base: \`${result.baseBranch}\``,
-      `- resolve: \`${renderResolveCommand(result.fix.resolveCommand)}\``,
-    ].join("\n"),
-  );
-
-  sections.push("## Instructions");
-  sections.push(result.fix.instructions.map((inst, i) => `${i + 1}. ${inst}`).join("\n"));
-
-  return sections.join("\n\n");
-}
-
-function blockquote(body: string): string {
-  return body
-    .split("\n")
-    .map((line) => (line === "" ? ">" : `> ${line}`))
-    .join("\n");
 }
