+++ src/cli/args.mts
@@ -5,9 +5,7 @@
 
 import { parseArgs } from "node:util";
 import { loadConfig } from "../config/load.mts";
-import type { GlobalOptions, ShepherdAction } from "../types.mts";
-import { deriveVerdict } from "../commands/status.mts";
-import type { PrSummary } from "../commands/status.mts";
+import type { GlobalOptions } from "../types.mts";
 
 // Flags that consume the next argument as their value (used for PR-number
 // detection only — prevents a flag's value from being mistaken for a PR number).
@@ -160,50 +158,10 @@ export function parseStatusPrNumbers(args: string[]): number[] {
   return prNumbers;
 }
 
-// ---------------------------------------------------------------------------
-// Duration parsing
-// ---------------------------------------------------------------------------
-
-export function parseDurationToMinutes(s: string, defaultMinutes?: number): number {
-  const m = /^(\d+)(m|min|minutes?|h|hours?)?$/.exec(s.trim());
-  if (!m) return defaultMinutes ?? loadConfig().watch.readyDelayMinutes;
-  const n = parseInt(m[1]!, 10);
-  const unit = m[2] ?? "m";
-  if (unit.startsWith("h")) return n * 60;
-  return n;
-}
-
-// ---------------------------------------------------------------------------
-// Exit code mapping
-// ---------------------------------------------------------------------------
-
-export function statusToExitCode(status: string): number {
-  switch (status) {
-    case "READY":
-      return 0;
-    case "IN_PROGRESS":
-      return 2;
-    case "UNRESOLVED_COMMENTS":
-      return 3;
-    default:
-      return 1;
-  }
-}
-
-export function iterateActionToExitCode(action: ShepherdAction): number {
-  switch (action) {
-    case "fix_code":
-    case "rebase":
-      return 1;
-    case "cancel":
-      return 2;
-    case "escalate":
-      return 3;
-    default:
-      return 0;
-  }
-}
-
-export function deriveSimpleReady(s: PrSummary): boolean {
-  return deriveVerdict(s) === "READY";
-}
+// Re-export exit-code helpers so existing importers of args.mts keep working.
+export {
+  parseDurationToMinutes,
+  statusToExitCode,
+  iterateActionToExitCode,
+  deriveSimpleReady,
+} from "./exit-codes.mts";
