+++ src/commands/check-status.mts
@@ -1,0 +1,44 @@
+import type { MergeStatusResult, ShepherdStatus } from "../types.mts";
+import type { getCiVerdict } from "../checks/classify.mts";
+
+export function computeStatus(
+  verdict: ReturnType<typeof getCiVerdict>,
+  unresolvedThreads: number,
+  unresolvedComments: number,
+  mergeStatus: MergeStatusResult,
+  changesRequestedReviews: number,
+): ShepherdStatus {
+  // Merge conflicts are always terminal regardless of CI state.
+  if (mergeStatus.status === "CONFLICTS") return "FAILING";
+  // Check CI state before merge-blocking states: BLOCKED/UNSTABLE/BEHIND are
+  // often caused by CI not having passed yet, so they shouldn't mask IN_PROGRESS.
+  // These merge-blocking states become PENDING (not FAILING) once CI is resolved.
+  if (verdict.anyFailing) return "FAILING";
+  if (verdict.anyInProgress) return "IN_PROGRESS";
+  // BLOCKED solely because a human reviewer hasn't approved yet — shepherd is done, hand off.
+  // copilotReviewInProgress means a bot still owes a review, which is not this case.
+  if (
+    verdict.allPassed &&
+    unresolvedThreads === 0 &&
+    unresolvedComments === 0 &&
+    changesRequestedReviews === 0 &&
+    mergeStatus.status === "BLOCKED" &&
+    !mergeStatus.copilotReviewInProgress &&
+    mergeStatus.reviewDecision === "REVIEW_REQUIRED"
+  ) {
+    return "READY";
+  }
+  if (
+    mergeStatus.status === "BLOCKED" ||
+    mergeStatus.status === "UNSTABLE" ||
+    mergeStatus.status === "BEHIND"
+  )
+    return "PENDING";
+  if (mergeStatus.status === "UNKNOWN") return "UNKNOWN";
+  if (changesRequestedReviews > 0) return "UNRESOLVED_COMMENTS";
+  if (unresolvedThreads > 0 || unresolvedComments > 0) return "UNRESOLVED_COMMENTS";
+  // DRAFT is treated the same as CLEAN for readiness — marking the PR ready resolves it.
+  if ((mergeStatus.status === "CLEAN" || mergeStatus.status === "DRAFT") && verdict.allPassed)
+    return "READY";
+  return "UNKNOWN";
+}
