+++ src/github/batch.mts
@@ -10,15 +10,9 @@
 import { graphql, graphqlWithRateLimit, type RateLimitInfo, type RepoInfo } from "./client.mts";
 import { paginateForward, paginateBackward } from "./pagination.mts";
 import { BATCH_PR_QUERY } from "./queries.mts";
-import type {
-  BatchPrData,
-  CheckConclusion,
-  CheckRun,
-  CheckStatus,
-  PrComment,
-  Review,
-  ReviewThread,
-} from "../types.mts";
+import { parseRawPr } from "./batch-parsers.mts";
+import type { RawBatchResponse, RawThread, RawComment, RawReview, RawReviewSummary, RawContextNode } from "./batch-raw-types.mts";
+import type { BatchPrData } from "../types.mts";
 
 // ---------------------------------------------------------------------------
 // Public API
@@ -189,275 +183,3 @@ export async function fetchPrBatch(
   return { data, rateLimit: result.rateLimit };
 }
 
-// ---------------------------------------------------------------------------
-// Parsers
-// ---------------------------------------------------------------------------
-
-function parseRawPr(
-  raw: RawPr,
-  rawThreadPages: RawThread[],
-  rawCommentNodes: RawComment[],
-  rawReviewNodes: RawReview[],
-  rawReviewSummaryNodes: RawReviewSummary[],
-  rawApprovedReviewNodes: RawReviewSummary[],
-  rawCheckNodes: RawContextNode[],
-): BatchPrData {
-  const reviewRequests = (raw.reviewRequests?.nodes ?? []).flatMap((n) => {
-    const login = n.requestedReviewer?.login ?? n.requestedReviewer?.name;
-    return login ? [{ login }] : [];
-  });
-
-  const latestReviews = (raw.latestReviews?.nodes ?? []).map((n) => ({
-    login: n.author?.login ?? "unknown",
-    state: n.state,
-  }));
-
-  const reviewThreads: ReviewThread[] = rawThreadPages.map((t) => {
-    const comment = t.comments.nodes[0];
-    return {
-      id: t.id,
-      isResolved: t.isResolved,
-      isOutdated: t.isOutdated,
-      isMinimized: comment?.isMinimized ?? false,
-      path: comment?.path ?? null,
-      line: comment?.line ?? null,
-      startLine: comment?.startLine ?? null,
-      author: comment?.author?.login ?? "unknown",
-      body: comment?.body ?? "",
-      createdAtUnix: comment ? parseCreatedAt(comment.createdAt) : 0,
-    };
-  });
-
-  const comments: PrComment[] = rawCommentNodes.map((c) => ({
-    id: c.id,
-    isMinimized: c.isMinimized,
-    author: c.author?.login ?? "unknown",
-    body: c.body,
-    createdAtUnix: parseCreatedAt(c.createdAt),
-  }));
-
-  const changesRequestedReviews: Review[] = rawReviewNodes.map((r) => ({
-    id: r.id,
-    author: r.author?.login ?? "unknown",
-    body: r.body,
-  }));
-
-  const reviewSummaries: Review[] = rawReviewSummaryNodes
-    .filter((r) => !r.isMinimized && r.body.trim() !== "")
-    .map((r) => ({
-      id: r.id,
-      author: r.author?.login ?? "unknown",
-      body: r.body,
-    }));
-
-  // APPROVED reviews often have empty bodies (clicking "Approve" without a comment), so
-  // we keep them — only the isMinimized filter applies. Monitor/iterate uses these IDs
-  // when the user opts in to minimizing approvals.
-  const approvedReviews: Review[] = rawApprovedReviewNodes
-    .filter((r) => !r.isMinimized)
-    .map((r) => ({
-      id: r.id,
-      author: r.author?.login ?? "unknown",
-      body: r.body,
-    }));
-
-  const checks: CheckRun[] = rawCheckNodes.flatMap((node) => {
-    if (node.__typename === "CheckRun") {
-      const event = node.checkSuite?.workflowRun?.event ?? null;
-      const runId = extractRunId(node.detailsUrl);
-      return [
-        {
-          name: node.name,
-          status: node.status as CheckRun["status"],
-          conclusion: node.conclusion as CheckRun["conclusion"],
-          detailsUrl: node.detailsUrl ?? "",
-          event,
-          runId,
-        },
-      ];
-    }
-    if (node.__typename === "StatusContext") {
-      const { status, conclusion } = mapStatusContextState(node.state);
-      return [
-        {
-          name: node.context,
-          status,
-          conclusion,
-          detailsUrl: node.targetUrl ?? "",
-          event: null,
-          runId: null,
-        },
-      ];
-    }
-    return [];
-  });
-
-  return {
-    nodeId: raw.id,
-    number: raw.number,
-    state: raw.state as BatchPrData["state"],
-    isDraft: raw.isDraft,
-    mergeable: raw.mergeable as BatchPrData["mergeable"],
-    mergeStateStatus: raw.mergeStateStatus as BatchPrData["mergeStateStatus"],
-    reviewDecision: (raw.reviewDecision ?? null) as BatchPrData["reviewDecision"],
-    headRefOid: raw.headRefOid,
-    baseRefName: raw.baseRefName,
-    reviewRequests,
-    latestReviews,
-    reviewThreads,
-    comments,
-    changesRequestedReviews,
-    reviewSummaries,
-    approvedReviews,
-    checks,
-  };
-}
-
-function parseCreatedAt(iso: string): number {
-  const ms = new Date(iso).getTime();
-  return Number.isFinite(ms) ? Math.floor(ms / 1000) : 0;
-}
-
-function extractRunId(url: string | undefined | null): string | null {
-  if (!url) return null;
-  const m = /\/runs\/(\d+)/.exec(url);
-  return m ? (m[1] ?? null) : null;
-}
-
-/** Maps a GitHub commit status `state` to CheckRun-compatible status + conclusion. */
-function mapStatusContextState(state: string): {
-  status: CheckStatus;
-  conclusion: CheckConclusion;
-} {
-  switch (state) {
-    case "SUCCESS":
-      return { status: "COMPLETED", conclusion: "SUCCESS" };
-    case "FAILURE":
-    case "ERROR":
-      return { status: "COMPLETED", conclusion: "FAILURE" };
-    case "PENDING":
-    case "EXPECTED":
-    default:
-      return { status: "IN_PROGRESS", conclusion: null };
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Raw GraphQL response types (private to this module)
-// ---------------------------------------------------------------------------
-
-interface RawBatchResponse {
-  repository: {
-    pullRequest: RawPr | null;
-  };
-}
-
-interface RawPr {
-  id: string;
-  number: number;
-  state: string;
-  isDraft: boolean;
-  mergeable: string;
-  mergeStateStatus: string;
-  reviewDecision: string | null;
-  headRefOid: string;
-  baseRefName: string;
-  reviewRequests: {
-    nodes: Array<{
-      requestedReviewer: { login?: string; name?: string } | null;
-    }>;
-  };
-  latestReviews: {
-    nodes: Array<{
-      author: { login: string } | null;
-      state: string;
-    }>;
-  };
-  reviewThreads: {
-    pageInfo: { hasPreviousPage: boolean; startCursor: string | null };
-    nodes: RawThread[];
-  };
-  comments: {
-    pageInfo: { hasPreviousPage: boolean; startCursor: string | null };
-    nodes: RawComment[];
-  };
-  changesRequestedReviews: {
-    pageInfo: { hasPreviousPage: boolean; startCursor: string | null };
-    nodes: RawReview[];
-  };
-  reviewSummaries: {
-    pageInfo: { hasPreviousPage: boolean; startCursor: string | null };
-    nodes: RawReviewSummary[];
-  };
-  approvedReviews: {
-    pageInfo: { hasPreviousPage: boolean; startCursor: string | null };
-    nodes: RawReviewSummary[];
-  };
-  commits: {
-    nodes: Array<{
-      commit: {
-        statusCheckRollup: {
-          contexts: {
-            pageInfo: { hasNextPage: boolean; endCursor: string | null };
-            nodes: RawContextNode[];
-          };
-        } | null;
-      };
-    }>;
-  };
-}
-
-interface RawThread {
-  id: string;
-  isResolved: boolean;
-  isOutdated: boolean;
-  comments: {
-    nodes: Array<{
-      id: string;
-      isMinimized: boolean;
-      author: { login: string } | null;
-      body: string;
-      path: string | null;
-      line: number | null;
-      startLine: number | null;
-      createdAt: string;
-    }>;
-  };
-}
-
-interface RawComment {
-  id: string;
-  isMinimized: boolean;
-  author: { login: string } | null;
-  body: string;
-  createdAt: string;
-}
-
-interface RawReview {
-  id: string;
-  author: { login: string } | null;
-  body: string;
-}
-
-interface RawReviewSummary {
-  id: string;
-  isMinimized: boolean;
-  author: { login: string } | null;
-  body: string;
-}
-
-type RawContextNode =
-  | {
-      __typename: "CheckRun";
-      name: string;
-      status: string;
-      conclusion: string | null;
-      detailsUrl: string | null;
-      checkSuite: { workflowRun: { event: string } | null } | null;
-    }
-  | {
-      __typename: "StatusContext";
-      context: string;
-      state: string;
-      targetUrl: string | null;
-    };
